some design changes and final commit before push

+++ src/Components/Customer/myApplied.js
@@ -1,8 +1,9 @@
 import React, { Component, useState } from 'react';
 import { Container, Row, Col } from 'react-bootstrap';
 import { Route, Link } from 'react-router-dom';
-import { Button } from "../../Header/Button";
-import { Card } from 'react-bootstrap';
+
+import { Card,Button } from 'react-bootstrap';
+import Swal from 'sweetalert2'
 //import { axios } from 'axios';
 const axios = require('axios').default;
 
@@ -18,7 +19,7 @@ class myApplied extends Component {
 
         axios({
             method: 'get',
-            url: 'http://localhost:91/job/showMyApplied',
+            url: 'http://localhost:91/job/showMyApplied2',
             headers: { 'authorization': `Bearer ${localStorage.getItem('token')}` }
         })
             .then((response) => {
@@ -37,16 +38,47 @@ class myApplied extends Component {
 
 
     deleteJob = (id) => {
-        axios.delete('http://localhost:91/job/deleteMyApplied/' + id, this.state.config)
-            .then((response) => {
-                console.log(response)
-                alert("Delete successfull")
-                window.location.reload()
-            })
-            .catch((err) => {
-                //console.log(err.response)
-                alert("Delete unsuccessfull")
-            })
+
+
+
+        Swal.fire({
+            title: 'Delete this application?',
+            icon: 'warning',
+      
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, Delete'
+          }).then((result) => {
+            if (result.isConfirmed) {
+      
+                axios.delete('http://localhost:91/job/deleteMyApplied/' + id, this.state.config)
+                .then((response) => {
+                    console.log(response)
+                    //alert("Delete successfull")
+                    Swal.fire(
+                        'Deleted',
+                        'Your have successfully deleted this application.',
+                        'success'
+                      )
+                    window.location.reload()
+                })
+                .catch((err) => {
+                    //console.log(err.response)
+                    alert("Delete unsuccessfull")
+                })
+              
+            }
+      
+          })
+      
+
+
+
+
+
+
+
+     
 
     }
 
@@ -77,44 +109,44 @@ class myApplied extends Component {
                         {
                             this.state.appliedjobs.map((job) => {
                                 return (<div>
-                                    <Card body>
-                                        <div class="card-text-center" >
-                                            <div class="card-header">
-                                                <h5 class="card-title">{job.jobid.jobtitle}</h5>
+                                    <div class="py-5 service-22">
+                                        <div class="container">
+
+                                            <div class="row wrap-service-22">
+
+                                                <div class="col-lg-6">
+                                                    <img src={`http://localhost:91/${job.jobid.photo}`} class="rounded img-shadow img-fluid" alt="wrapkit" style={{ height: "400px" }} />
+                                                </div>
+
+                                                <div class="col-lg-6 mt-4 mt-md-0">
+                                                    <div class="text-box">
+                                                        <h4 class="font-weight-light mt-2 mb-4">Job Title: <span class="text-megna">{job.jobid.jobtitle}</span></h4>
+                                                        <h6 class="font-weight-light">Job Type: <span class="text-megna">{job.jobid.jobtype}</span></h6>
+                                                        <h6 class="font-weight-light">Description: </h6><p>{job.jobid.jobdescription}</p>
+
+                                                        <h6 class="font-weight-light">Required Experience: <span class="text-megna">{job.jobid.requiredexperience}</span></h6>
+                                                        <h6 class="font-weight-light">Job Price: <span class="text-megna">{job.jobid.jobprice}</span></h6>
+                                                        {/* <h6 class="font-weight-light">Email of company: <span class="text-megna">{job.jobid.creator.email}</span></h6>
+                                                        <h6 class="font-weight-light">Company: <span class="text-megna">{job.jobid.creator.company}</span></h6> */}
+                                                        <h6 class="font-weight-light">Applied on: <span class="text-megna">{job.createdAt}</span></h6>
+
+                                                        {
+                                                            job.confirmStatus === "Confirmed"
+                                                                ? (<p>Your job application has been confirmed by company.</p>)
+                                                                : job.confirmStatus === "denied"
+                                                                    ? (<p>Your job application has been denied by the company.</p>)
+
+                                                                    : (<p class=".text-primary">Your job application is not reviewed by company.</p>)
+                                                        }
+
+                                                        <button class="btn btn-info-gradiant btn-md text-white border-0" onClick={this.deleteJob.bind(this, job._id)}><span>Delete this application</span></button>
+
+                                                    </div>
+                                                </div>
+
                                             </div>
-                                            <div class="card-body">
-                                            <img src={`http://localhost:91/${job.jobid.photo}`} className="img-fluid" style={{ height: "400px" }} />
-                                                <p class="card-text-center">Type :{job.jobid.jobtype}</p>
-                                                <p>Description:{job.jobid.jobdescription}</p>
-                                                <p>Required experience:{job.jobid.requiredexperience}</p>
-                                                <p>Email:{job.jobid.email}</p>
-                                                <p>Company:{job.jobid.company}</p>
-                                                <p>Posted at:{job.createdAt}</p>
-
-                                                {
-                                                    job.confirmStatus === "Confirmed"
-                                                        ? (<p>Your job application has been confirmed by company.</p>)
-                                                        : job.confirmStatus === "denied"
-                                                        ? (<p>Your job application has been denied by the company.</p>)
-
-                                                        : (<p>Your job application is not reviewed by company.</p>)
-                                                }
-
-                                                <p><Button onClick={this.deleteJob.bind(this, job._id)}>Delete</Button></p>
-
-
-                                            </div>
-
                                         </div>
-
-
-
-
-
-
-
-                                    </Card>
-                                    <p></p>
+                                    </div>
                                 </div>
                                 )
                             })
